kubevirt: move ResourceLoadErrors to its own directory
